# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "2.0.0",
    "title": "Helpoo Api Documentation",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module, for <b>Helpoo</b>."
  },
  "servers": [
    {
      "url": "http://apidev.helpooapp.net/"
    },
    {
      "url": "https://apidev.helpooapp.net/"
    }
  ],
...
  "paths": {
...
    "/api/v2/accidentTypes/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/accidentTypes/images": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accidentTypesIds": {
                    "example": "any"
                  },
                  "inspection": {
                    "example": "any"
                  },
                  "inspectionId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/clients/getAll": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/clients/getOne/{clientId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v2/clients/edit/{clientId}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "edit client",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "active": {
                  "type": "string",
                  "example": "true"
                },
                "confirmed": {
                  "type": "string",
                  "example": "true"
                }
              },
              "required": [
                "active",
                "confirmed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v2/serviceRequest/types": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authentication",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}